Make changes as requested
